api_service/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=4.6
psycopg2-binary>=2.9
pydantic[email]>=1.10
bcrypt>=4.0.1
python-multipart>=0.0.5
python-dotenv>=0.19.0